Added stuff to Edit and main and killed bugs
Main Project:
Added a proper FPS timer for Updates in the main project.
LevelEditor:
Added so that clicking a laser in LevelEditor selects it's holder.
Other:
Seperated XmlSaver into XmlSaver and LevelSaver. Added MiniBots to
LevelSaver and fixed other bugs with it.
Fixed so Laserholder's setPosition also sets the Laser's position. Fixed
UnitWrap Bug with animation.